Move document colors to their own modules
